[median] Исправил типы данных

+++ src/utils/image/Image.h
@@ -7,7 +7,7 @@ class Image
 {
 public:
 	explicit Image(const std::string& path);
-	Image(unsigned int width, unsigned int height, unsigned int channels, const unsigned char* data);
+	Image(uint32_t width, uint32_t height, uint8_t channels, const unsigned char* data);
 	~Image();
 
 	Image(const Image&) = delete;
@@ -19,10 +19,10 @@ public:
 
 	bool IsRGB() const;
 
-	unsigned int GetWidth() const;
-	unsigned int GetHeight() const;
-	unsigned int GetChannels() const;
-	size_t GetPixelCount() const;
+	uint32_t GetWidth() const;
+	uint32_t GetHeight() const;
+	uint8_t GetChannels() const;
+	uint64_t GetPixelCount() const;
 
 	const unsigned char* GetData() const;
 	unsigned char* GetData();
@@ -32,8 +32,8 @@ private:
 	void Free() noexcept;
 
 private:
-	unsigned int m_width = 0;
-	unsigned int m_height = 0;
-	unsigned int m_channels = 0;
+	uint32_t m_width = 0;
+	uint32_t m_height = 0;
+	uint8_t m_channels = 0;
 	unsigned char* m_data = nullptr;
 };
